code review: 훅 분리 및 네이밍 개선
