+++ app/sign-up/(profile)/_layout.tsx
@@ -1,29 +1,20 @@
-import { router, Stack } from "expo-router";
-import { useState } from "react";
+import { Stack } from "expo-router";
 
 import { SignUpHeader } from "@screens/sign-up/components";
-import { NicknameGenreContext } from "@screens/sign-up/hooks";
+import { NicknameGenreContext, useSignUpProfile } from "@screens/sign-up/hooks";
 import { colors } from "@theme/token";
 
 export default function NicknameGenreLayout() {
-  // TODO: 훅으로 통합 => (screens/sign-up/nickname)
-  const [nickname, setNickname] = useState("");
-  const [isNicknameDuplicated, setIsNicknameDuplicated] = useState(false);
-  const [genre, setGenre] = useState<string | null>(null);
-
-  const disabledNickname = nickname.trim().length < 2;
-
-  // TODO: 중복 여부는 추후 서버 api와 연동하여 판별. 성공 시 이동
-  const handleToSelectGenre = () => {
-    setIsNicknameDuplicated(false);
-    router.push("/sign-up/genre");
-  };
-
-  // TODO: 서버에 회원가입 요청
-  const handleToOnboarding = () => {
-    alert(`닉네임: ${nickname} / 장르: ${genre}`);
-    router.push("/sign-up/onboarding");
-  };
+  const {
+    nickname,
+    setNickname,
+    isNicknameDuplicated,
+    genre,
+    setGenre,
+    disabledNickname,
+    handleToSelectGenre,
+    handleToOnboarding,
+  } = useSignUpProfile();
 
   return (
     <NicknameGenreContext.Provider
