+++ screens/sign-up/hooks/index.ts
@@ -2,3 +2,4 @@ export {
   NicknameGenreContext,
   useNicknameGenreContext,
 } from "./use-nickname-genre-context";
+export { useSignUpProfile } from "./use-sign-up-profile";
