+++ screens/sign-up/nickname/index.ts
@@ -1,2 +1,2 @@
-export { InputNickname, SignUpNicknameHeader } from "./components";
+export { InputNickname } from "./components";
 export { default as SignUpNicknameScreen } from "./sign-up-nickname-screen";
