+++ src/shared/constants/regex.ts
@@ -1,1 +1,1 @@
-export const ALLOWED_NICKNAME_REGEX = /[^0-9a-zㄱ-ㅎㅏ-ㅣ가-힣]/g;
+export const INVALID_NICKNAME_CHARS_REGEX = /[^0-9a-zㄱ-ㅎㅏ-ㅣ가-힣]/g;
